integrated completion and original functionality
