--- lib/Extension/ClassToFile/ClassToFileExtension.php
@@ -1,80 +1,0 @@
-<?php
-
-namespace Phpactor\Extension\ClassToFile;
-
-use Phpactor\Container\Extension;
-use Phpactor\Container\ContainerBuilder;
-use Phpactor\MapResolver\Resolver;
-use Phpactor\Extension\ClassToFile\Command\FileInfoCommand;
-use Phpactor\ClassFileConverter\Domain\ClassToFileFileToClass;
-use Phpactor\ClassFileConverter\Adapter\Composer\ComposerClassToFile;
-use Phpactor\ClassFileConverter\Adapter\Simple\SimpleClassToFile;
-use Phpactor\ClassFileConverter\Domain\ChainClassToFile;
-use Phpactor\ClassFileConverter\Adapter\Composer\ComposerFileToClass;
-use Phpactor\ClassFileConverter\Adapter\Simple\SimpleFileToClass;
-use Phpactor\Extension\Core\CoreExtension;
-use Phpactor\Container\Container;
-use Phpactor\Extension\ClassToFile\Application\FileInfo;
-use Phpactor\ClassFileConverter\Domain\ChainFileToClass;
-
-class ClassToFileExtension implements Extension
-{
-    /**
-     * {@inheritDoc}
-     */
-    public function load(ContainerBuilder $container)
-    {
-        $container->register('command.file_info', function (Container $container) {
-            return new FileInfoCommand(
-                $container->get('application.file_info'),
-                $container->get('console.dumper_registry')
-            );
-        }, [ 'ui.console.command' => [ 'name' => 'file:info' ]]);
-        $container->register('class_to_file.converter', function (Container $container) {
-            return new ClassToFileFileToClass(
-                $container->get('class_to_file.class_to_file'),
-                $container->get('class_to_file.file_to_class')
-            );
-        });
-
-        $container->register('class_to_file.class_to_file', function (Container $container) {
-            $classToFiles = [];
-            foreach ($container->get('composer.class_loaders') as $classLoader) {
-                $classToFiles[] = new ComposerClassToFile($classLoader);
-            }
-
-            if (empty($classToFiles)) {
-                $classToFiles[] = new SimpleClassToFile($container->getParameter(CoreExtension::WORKING_DIRECTORY));
-            }
-
-            return new ChainClassToFile($classToFiles);
-        });
-
-        $container->register('class_to_file.file_to_class', function (Container $container) {
-            $fileToClasses = [];
-            foreach ($container->get('composer.class_loaders') as $classLoader) {
-                $fileToClasses[] =  new ComposerFileToClass($classLoader);
-            }
-
-            if (empty($fileToClasses)) {
-                $fileToClasses[] = new SimpleFileToClass();
-            }
-
-            return new ChainFileToClass($fileToClasses);
-        });
-
-        $container->register('application.file_info', function (Container $container) {
-            return new FileInfo(
-                $container->get('class_to_file.converter'),
-                $container->get('source_code_filesystem.simple')
-            );
-        });
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function configure(Resolver $schema)
-    {
-    }
-}
