+++ lib/Extension/CodeTransform/CodeTransformExtension.php
@@ -29,6 +29,7 @@ use Phpactor\Config\Paths;
 use Phpactor\Container\Container;
 use Phpactor\Container\ContainerBuilder;
 use Phpactor\Container\Extension;
+use Phpactor\Extension\WorseReflection\WorseReflectionExtension;
 use Phpactor\MapResolver\Resolver;
 use Phpactor\Extension\CodeTransform\Application\ClassInflect;
 use Phpactor\Extension\CodeTransform\Application\ClassNew;
@@ -153,14 +154,14 @@ class CodeTransformExtension implements Extension
 
         $container->register('code_transform.transformer.complete_constructor', function (Container $container) {
             return new CompleteConstructor(
-                $container->get('reflection.reflector'),
+                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                 $container->get('code_transform.updater')
             );
         }, [ 'code_transform.transformer' => [ 'name' => 'complete_constructor' ]]);
 
         $container->register('code_transform.transformer.implement_contracts', function (Container $container) {
             return new ImplementContracts(
-                $container->get('reflection.reflector'),
+                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                 $container->get('code_transform.updater'),
                 $container->get('code_transform.builder_factory')
             );
@@ -174,7 +175,7 @@ class CodeTransformExtension implements Extension
 
         $container->register('code_transform.transformer.add_missing_properties', function (Container $container) {
             return new AddMissingProperties(
-                $container->get('reflection.reflector'),
+                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                 $container->get('code_transform.updater')
             );
         }, [ 'code_transform.transformer' => [ 'name' => 'add_missing_properties' ]]);
@@ -196,7 +197,7 @@ class CodeTransformExtension implements Extension
         $container->register('code_transform.from_existing_generators', function (Container $container) {
             $generators = [
                 'interface' => new InterfaceFromExistingGenerator(
-                    $container->get('reflection.reflector'),
+                    $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                     $container->get('code_transform.renderer')
                 ),
             ];
@@ -237,14 +238,14 @@ class CodeTransformExtension implements Extension
     {
         $container->register('code_transform.refactor.extract_constant', function (Container $container) {
             return new WorseExtractConstant(
-                $container->get('reflection.reflector'),
+                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                 $container->get('code_transform.updater')
             );
         });
 
         $container->register('code_transform.refactor.generate_method', function (Container $container) {
             return new WorseGenerateMethod(
-                $container->get('reflection.reflector'),
+                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                 $container->get('code_transform.builder_factory'),
                 $container->get('code_transform.updater')
             );
@@ -252,7 +253,7 @@ class CodeTransformExtension implements Extension
 
         $container->register('code_transform.refactor.generate_accessor', function (Container $container) {
             return new WorseGenerateAccessor(
-                $container->get('reflection.reflector'),
+                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                 $container->get('code_transform.updater'),
                 $container->getParameter(self::GENERATE_ACCESSOR_PREFIX),
                 $container->getParameter(self::GENERATE_ACCESSOR_UPPER_CASE_FIRST)
@@ -265,7 +266,7 @@ class CodeTransformExtension implements Extension
 
         $container->register('code_transform.refactor.override_method', function (Container $container) {
             return new WorseOverrideMethod(
-                $container->get('reflection.reflector'),
+                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                 $container->get('code_transform.builder_factory'),
                 $container->get('code_transform.updater')
             );
@@ -273,7 +274,7 @@ class CodeTransformExtension implements Extension
 
         $container->register('code_transform.refactor.extract_method', function (Container $container) {
             return new WorseExtractMethod(
-                $container->get('reflection.reflector'),
+                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                 $container->get('code_transform.builder_factory'),
                 $container->get('code_transform.updater')
             );
@@ -303,7 +304,7 @@ class CodeTransformExtension implements Extension
             );
         });
         $container->register('code_transform.builder_factory', function (Container $container) {
-            return new WorseBuilderFactory($container->get('reflection.reflector'));
+            return new WorseBuilderFactory($container->get(WorseReflectionExtension::SERVICE_REFLECTOR));
         });
     }
 
@@ -367,7 +368,7 @@ class CodeTransformExtension implements Extension
 
         $container->register('rpc.handler.override_method', function (Container $container) {
             return new OverrideMethodHandler(
-                $container->get('reflection.reflector'),
+                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR),
                 $container->get('code_transform.refactor.override_method')
             );
         }, [ 'rpc.handler' => [] ]);
