+++ lib/Extension/CompletionExtra/Command/CompleteCommand.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Phpactor\Extension\Completion\Command;
+namespace Phpactor\Extension\CompletionExtra\Command;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Phpactor\Extension\Core\Console\Dumper\DumperRegistry;
-use Phpactor\Extension\Completion\Application\Complete;
+use Phpactor\Extension\CompletionExtra\Application\Complete;
 use Phpactor\Extension\Core\Application\Helper\FilesystemHelper;
 use Phpactor\Extension\Core\Console\Handler\FormatHandler;
