+++ lib/Extension/CompletionExtra/CompletionExtraExtension.php
@@ -1,0 +1,85 @@
+<?php
+
+namespace Phpactor\Extension\CompletionExtra;
+
+use Phpactor\Completion\Bridge\TolerantParser\SourceCodeFilesystem\ScfClassCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseClassAliasCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseConstantCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseConstructorCompletor;
+use Phpactor\Completion\Bridge\WorseReflection\Formatter\FunctionFormatter;
+use Phpactor\Completion\Bridge\WorseReflection\Formatter\ClassFormatter;
+use Phpactor\Completion\Bridge\WorseReflection\Formatter\MethodFormatter;
+use Phpactor\Completion\Bridge\WorseReflection\Formatter\ParametersFormatter;
+use Phpactor\Completion\Bridge\WorseReflection\Formatter\VariableFormatter;
+use Phpactor\Completion\Bridge\TolerantParser\ChainTolerantCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseClassMemberCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseFunctionCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseParameterCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseLocalVariableCompletor;
+use Phpactor\Completion\Core\Formatter\ObjectFormatter;
+use Phpactor\Completion\Bridge\WorseReflection\Formatter\ParameterFormatter;
+use Phpactor\Completion\Bridge\WorseReflection\Formatter\PropertyFormatter;
+use Phpactor\Completion\Bridge\WorseReflection\Formatter\TypeFormatter;
+use Phpactor\Completion\Bridge\WorseReflection\Formatter\TypesFormatter;
+use Phpactor\Completion\Core\ChainCompletor;
+use Phpactor\Container\Extension;
+use Phpactor\Container\ContainerBuilder;
+use Phpactor\Extension\CompletionExtra\Rpc\CompleteHandler;
+use Phpactor\Extension\CompletionExtra\Rpc\HoverHandler;
+use Phpactor\MapResolver\Resolver;
+use Phpactor\Container\Container;
+use Phpactor\Extension\CompletionExtra\Command\CompleteCommand;
+use Phpactor\Extension\CompletionExtra\Application\Complete;
+use Phpactor\Extension\CompletionExtra\LanguageServer\CompletionLanguageExtension;
+
+class CompletionExtraExtension implements Extension
+{
+    const CLASS_COMPLETOR_LIMIT = 'completion.completor.class.limit';
+
+    /**
+     * {@inheritDoc}
+     */
+    public function load(ContainerBuilder $container)
+    {
+        $this->registerCommands($container);
+        $this->registerLanguageServer($container);
+        $this->registerApplicationServices($container);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function configure(Resolver $schema)
+    {
+    }
+
+    private function registerCommands(ContainerBuilder $container)
+    {
+        $container->register('command.complete', function (Container $container) {
+            return new CompleteCommand(
+                $container->get('application.complete'),
+                $container->get('console.dumper_registry')
+            );
+        }, [ 'ui.console.command' => []]);
+    }
+
+    private function registerApplicationServices(ContainerBuilder $container)
+    {
+        $container->register('application.complete', function (Container $container) {
+            return new Complete(
+                $container->get('completion.completor')
+            );
+        });
+    }
+
+    private function registerLanguageServer(ContainerBuilder $container)
+    {
+        $container->register('completion.language_server.completion', function (Container $container) {
+            return new CompletionLanguageExtension(
+                $container->get('language_server.session_manager'),
+                $container->get('completion.completor'),
+                $container->get('reflection.reflector')
+            );
+        }, [ 'language_server.extension' => [] ]);
+    }
+}
