+++ lib/Extension/CompletionExtra/LanguageServer/CompletionLanguageExtension.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Extension\Completion\LanguageServer;
+namespace Phpactor\Extension\CompletionExtra\LanguageServer;
 
 use LanguageServerProtocol\CompletionOptions;
 use LanguageServerProtocol\ServerCapabilities;
@@ -10,6 +10,7 @@ use Phpactor\LanguageServer\Core\Extension;
 use Phpactor\LanguageServer\Core\Handlers;
 use Phpactor\LanguageServer\Core\Session\Manager;
 use Phpactor\WorseReflection\Core\Reflector\SourceCodeReflector;
+use Phpactor\Extension\CompletionExtra\LanguageServer\CompletionHandler;
 
 class CompletionLanguageExtension implements Extension
 {
